refactor: #62 exception 수정 및 테스트 코드 수정
